Common deployment workflow (#324)
Merging the frontend and backend deployment workflows, and preparing the manual trigger workflow for when the autodetection on portainer is fixed.

This first gets the current datetime to use as a versioning tag (down to the minute). Launches the actual builds. Then if either frontend or backend has successfully uploader a new container package, a git tag is created.