Removed unused and outdated mockoon specs (#330)
We used to use mockoon to mock the backend back when it was incomplete. But it hasn't been updated in a long time. With a now complete backend, I doubt keeping and updating mockoon really offers much value over properly using and testing the backend.

It's also been the main reason for past high/critical vulnerabilities (due to transitive dependencies)

Removing `@mockoon/cli` has `removed 168 packages` and fixed 4 moderate vulnerability warnings as of this date.